cmsc491-vision/hw3-cat_dog/requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
